LayoutTests/fast/css/fontloader-multiple-faces-expected.txt

Test load events for a font family consists of multiple @font-faces.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS events['loading'] is 1
PASS events['loadstart'] is 2
PASS events['load'] is 2
PASS events['error'] is undefined
PASS events['loadingdone'] is undefined
PASS document.fonts.checkFont('10px TestFont') is true
PASS events['loadingdone'] is 1
PASS successfullyParsed is true

TEST COMPLETE
